+++ r5dev/vstdlib/callback.cpp
@@ -687,12 +687,12 @@ void VPK_Mount_f(const CCommand& args)
 		return;
 	}
 
-	if (g_pFullFileSystem)
+	if (g_pFullFileSystem) // Initialized when g_pFileSystem_Stdio is initialized (global engine pointer).
 	{
-		VPKData_t* pPakData = g_pFullFileSystem->MountVPK(args.Arg(1));
+		VPKData_t* pPakData = g_pFileSystem_Stdio->MountVPK(args.Arg(1));
 		if (pPakData)
 		{
-			DevMsg(eDLL_T::FS, "Mounted VPK file '%s' with handle '%d'\n", args.Arg(1), pPakData->m_nHandle);
+			DevMsg(eDLL_T::FS, "Mounted VPK file '%s' with handle '%i'\n", args.Arg(1), pPakData->m_nHandle);
 		}
 		else
 		{
@@ -701,7 +701,7 @@ void VPK_Mount_f(const CCommand& args)
 	}
 	else
 	{
-		Warning(eDLL_T::FS, "Unable to mount VPK file '%s': '%s' is not initalized\n", args.Arg(1), VAR_NAME(g_pFileSystem));
+		Warning(eDLL_T::FS, "Unable to mount VPK file '%s': '%s' is not initalized\n", args.Arg(1), VAR_NAME(g_pFullFileSystem));
 	}
 }
